requirements.txt
Flask==2.0.1
Flask-Caching==1.10.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Werkzeug==2.0.1
click==8.0.1
itsdangerous==2.0.1
nltk==3.6.3
numpy==2.0.0
pandas==2.2.2
praw==7.4.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
six==1.16.0
tqdm==4.62.2
update-checker==0.18.0
urllib3==1.26.6
gunicorn==20.1.0